Move typeclasses test
