+++ modules/core/arrow-data/src/test/kotlin/arrow/data/NonEmptyListTest.kt
@@ -5,6 +5,7 @@ import arrow.instances.nonemptylist.applicative.applicative
 import arrow.instances.nonemptylist.comonad.comonad
 import arrow.instances.nonemptylist.eq.eq
 import arrow.instances.nonemptylist.monad.monad
+import arrow.instances.nonemptylist.semigroup.semigroup
 import arrow.instances.nonemptylist.semigroupK.semigroupK
 import arrow.instances.nonemptylist.show.show
 import arrow.instances.nonemptylist.traverse.traverse
@@ -12,6 +13,7 @@ import arrow.test.UnitSpec
 import arrow.test.laws.*
 import arrow.typeclasses.Eq
 import io.kotlintest.KTestJUnitRunner
+import io.kotlintest.properties.forAll
 import org.junit.runner.RunWith
 
 @RunWith(KTestJUnitRunner::class)
@@ -28,8 +30,8 @@ class NonEmptyListTest : UnitSpec() {
         NonEmptyList.applicative(),
         Eq.any()),
       ComonadLaws.laws(NonEmptyList.comonad(), { NonEmptyList.of(it) }, Eq.any()),
-      TraverseLaws.laws(NonEmptyList.traverse(), NonEmptyList.applicative(), { n: Int -> NonEmptyList.of(n) }, Eq.any())
+      TraverseLaws.laws(NonEmptyList.traverse(), NonEmptyList.applicative(), { n: Int -> NonEmptyList.of(n) }, Eq.any()),
+      SemigroupLaws.laws(NonEmptyList.semigroup(), Nel(1, 2, 3), Nel(3, 4, 5), Nel(6, 7, 8), NonEmptyList.eq(Int.eq()))
     )
-
   }
 }
