+++ modules/core/arrow-typeclasses/src/test/kotlin/arrow/typeclasses/NumberMonoidTest.kt
@@ -1,4 +1,4 @@
-package arrow.data
+package arrow.typeclasses
 
 import arrow.instances.*
 import arrow.test.UnitSpec
